+++ include/parser.hpp
@@ -10,12 +10,12 @@
 class Parser
 {
 public:
-    void parse(int argc, char *argv[], bool &isRestoring, std::string &fileName, std::string &outputFileName, std::string &password, int &frameHeight, int &frameWidth, float &bitPixelRatio);
+    void parse(int argc, char *argv[], bool &isRestoring, std::string &fileName, std::string &outputFileName, std::string &password, int &frameHeight, int &frameWidth, int &bitPixelRatio);
 
 private:
     std::vector<std::string> params;
-    void validateArguments(bool &isRestoring, std::string &fileName, std::string &outputFileName, std::string &password, int &frameHeight, int &frameWidth, float &bitPixelRatio);
-    void prepareArguments(bool &isRestoring, std::string &fileName, std::string &outputFileName, std::string &password, int &frameHeight, int &frameWidth, float &bitPixelRatio);
+    void validateArguments(bool &isRestoring, std::string &fileName, std::string &outputFileName, std::string &password, int &frameHeight, int &frameWidth, int &bitPixelRatio);
+    void prepareArguments(bool &isRestoring, std::string &fileName, std::string &outputFileName, std::string &password, int &frameHeight, int &frameWidth, int &bitPixelRatio);
     void displayHelpTexts();
     friend void displayEnteredArguments(Parser *parser);
     friend void markErrorPart(int index, Parser *parser);
