+++ src/ziply.cpp
@@ -2,7 +2,7 @@
 #include "generator.hpp"
 
 
-Ziply::Ziply(const std::string &fileName, const std::string &outputFileName, const std::string &password, const int frameWidth, const int frameHeight, float bitPixelRatio)
+Ziply::Ziply(const std::string &fileName, const std::string &outputFileName, const std::string &password, const int frameWidth, const int frameHeight, int bitPixelRatio)
     : fileName(fileName), outputFileName(outputFileName), password(password), frameWidth(frameWidth), frameHeight(frameHeight), bitPixelRatio(bitPixelRatio) {}
 
 void Ziply::convert()
